backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bollettometro"
version = "0.1.0"
description = "Bollettometro 2030 - API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pydantic>=2",
    "pydantic-settings>=2",
    "python-multipart>=0.0.6",
    "redis>=5",
    "celery[redis]>=5.3",
    "openai>=1.0",
    "reportlab>=4.0",
    "qrcode[pil]>=7.4",
    "Pillow>=10",
    "pdfplumber>=0.10",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "httpx>=0.26",
    "ruff>=0.1",
    "black>=24",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
